+++ settings.gradle.kts
@@ -33,4 +33,3 @@ include(":feature-movies")
 include(":core-database")
 include(":navigation-contracts")
 include(":core-preferences")
-include(":di")
